First stab at handling errors
The basic idea is that default connections between nodes in the graph
define a 'happy path'. Exceptions and explicitly returned errors from
any transducer are available via the 'err' subnode or, if you want both
values and errors, using the 'all' subnode.

This has implications for sources and sinks too.