reductions: use initializer (not initialState)
Using an initializer function allows for resetting the state whenever
the transducer is called with a reducer. This prevents state leaking
from earlier uses of a reductions transducer to later instances.

Especially bad since I 'cleverly' rewrote most of the transducers to
rely on reductions and mapcat and they almost always mutated the state.

Also: change the behavior of multiplex when given zero transducers to be
dropAll instead of identity (aka, takeAll). This degenerate behavior is
makes more sense.