Refactor graph and walkGraph
This refactoring has two parts:

part 1: moves the work of declaring all parent/child connections into
the graph definition instead of building this up in a prewalk stage
during graph walking. This part  also moves the cycle check to be at
graph definition as well.

part 2: sources and sinks are now unified into the concept of edges
which allows for a coupling between a source and sink when declaring an
edge node. In effect, edge nodes 'look' like they can support a cycle in
the graph since they act both as a sink and a source. Under the hood, an
edge node is a subgraph with the sink and source nodes unconnected to
each other.

Also, add debug to util.js