Fix timing bug in spread transducer
Previously, spread would 'jump the gun' by calling RESULT on child
transducers as soon as each child transducer returned a reduced state.

Now, spread calls RESULT on child transducers _only_ when RESULT is
called on spread itself.

The effect is that RESULT may never be called on a child transducer if
the incoming stream of data never ends.

Another way to explain the issue is that adding a spread transducer
upstream of a transducer should not change the timing semantics of when
RESULT is called.