Remove `edge` nodes
Also, stop augmenting nodes (both source/sink and xf nodes). The purpose
of augmenting them was to capture errors bug it felt pretty bulky and
made the graphs hard to visualize. The new plan to handle errors is to
route uncaught errors to an 'err' source pipe (if one is created). This
is much less work.