Sigh. Revert timing changes to `spread`.
Reverted spread semantics back to sending an early RESULT call down to
reduced child transducers. I really though this was wrong since the
after transducer was not doing what I expected.

Also: change the semantics to `after` which is what I should have done
in the first place. Instead of 'dropping' all input, which casued it to
declare itself reduced, it now 'ignores' all input and patiently waits
for the input to end.